Implemented Testcases and moved into repective folder Removed Common Module
