+++ app/build.gradle
@@ -4,7 +4,6 @@ plugins {
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
 
-
 }
 
 android {
@@ -32,16 +31,16 @@ android {
     kotlinOptions {
         jvmTarget = '1.8'
     }
-    buildFeatures{
+    buildFeatures {
         dataBinding true
     }
 }
 
 dependencies {
-    testImplementation project(":common:core-testing")
+
     implementation project(':data')
     implementation project(path: ':domain')
-    def worker_hilt="1.0.0"
+    def worker_hilt = "1.0.0"
     def hilt = "2.40.4"
     def hilt_compiler = "2.38.1"
     //Hilt
@@ -55,7 +54,7 @@ dependencies {
 
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1"
     // LiveData
-    implementation"androidx.lifecycle:lifecycle-livedata-ktx:2.5.1"
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.5.1"
     //Coroutine
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4"
@@ -73,10 +72,10 @@ dependencies {
     testImplementation "org.mockito:mockito-inline:3.10.0"
     testImplementation 'androidx.arch.core:core-testing:2.1.0'
     //mock wb
-    testImplementation"com.squareup.okhttp3:mockwebserver:4.10.0"
+    testImplementation "com.squareup.okhttp3:mockwebserver:4.10.0"
     //Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation  "com.squareup.retrofit2:converter-gson:2.9.0"
+    implementation "com.squareup.retrofit2:converter-gson:2.9.0"
     //truth
     testImplementation "com.google.truth:truth:1.1"
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6.1"
