+++ app/src/test/java/com/example/mvvmcleanarchidemo1/presentation/viewmodel/GameFakeService.kt
@@ -1,10 +1,10 @@
-package com.example.core_testing
+package com.example.mvvmcleanarchidemo1.presentation.viewmodel
 
 import com.example.data.api.GamesApiService
 import com.example.data.models.GamesAPIResponseDTO
 
 
-class GameFakeService(private val gamesData: GamesAPIResponseDTO): GamesApiService {
+class GameFakeService(private val gamesData: GamesAPIResponseDTO) : GamesApiService {
     override suspend fun getGamesData(): GamesAPIResponseDTO {
         return gamesData
     }
