+++ domain/src/test/java/com/example/domain/usecase/FakeDataGenerator.kt
@@ -1,0 +1,9 @@
+package com.example.domain.usecase
+
+import com.example.domain.entity.Amiibo
+
+object FakeDataGenerator {
+
+    val amiiboList=listOf(Amiibo("Super Smash Bros","https://raw.githubusercontent.com/N3evin/AmiiboAPI/master/images/icon_00000000-00000002.png","Mario"))
+
+}
