+++ domain/src/test/java/com/example/domain/usecase/GamesUseCaseTest.kt
@@ -2,10 +2,6 @@ package com.example.domain.usecase
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.example.core_testing.FakeDataGenerator
-import com.example.core_testing.GameFakeService
-import com.example.data.repository.GamesRepoImpl
-import com.example.data.repository.datasourceimpl.RemoteDataSourceImpl
 import com.example.domain.repository.GamesRepo
 import com.google.common.truth.Truth
 import kotlinx.coroutines.runBlocking
@@ -20,12 +16,10 @@ class GamesUseCaseTest {
     var instantTaskExecutorRule = InstantTaskExecutorRule()
     private lateinit var gameRepository: GamesRepo
     private lateinit var subject: GetGamesUseCase
-    private lateinit var api: GameFakeService
 
     @Before
     fun setup() {
-        api = GameFakeService(FakeDataGenerator.amiibo)
-        gameRepository = GamesRepoImpl(RemoteDataSourceImpl(api))
+        gameRepository= FakeGamesRepo()
         subject = GetGamesUseCase(gameRepository)
     }
